database/database.py
from pathlib import Path
from decouple import config


BASE_DIR = Path(__file__).resolve().parent.parent

DATABASES = {
    "default":{
        'ENGINE': 'django.db.backends.postgresql',
        'NAME': config('DATABASE_NAME'),
        'USER': config('DATABASE_USER'),
        'PASSWORD': config('DATABASE_PASSWORD'),
        'HOST': config('DATABASE_HOST'),
        'PORT': config('DATABASE_PORT'),
    }
}