requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==5.0.2
django-allauth==0.57.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
ffmpeg-python==0.2.0
future==0.18.3
idna==3.6
iniconfig==2.0.0
kombu==5.3.5
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psutil==6.0.0
py==1.11.0
pycparser==2.21
pydub==0.25.1
PyJWT==2.8.0
pytest==8.2.2
pytest-celery==1.0.0
pytest-django==4.8.0
pytest-docker-tools==3.1.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytube==15.0.0
pytz==2024.1
pywin32==306
redis==5.0.2
requests==2.31.0
requests-mock==1.12.1
requests-oauthlib==1.3.1
retry==0.9.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
